bug fixed.

+++ server_side/moderators/moderators/init.lua
@@ -96,8 +96,8 @@ local check_time = ""
 
 minetest.register_globalstep(function(dtime)
 
-	local check_file = minetest.get_worldpath() .. "/check_file" --means reset privs for supervisors
-	local check_file_worldedit = minetest.get_worldpath() .. "/check_file_worldedit" --means add worldedit privs for supervisors
+	local check_file_reset = minetest.get_worldpath() .. "/check_file_reset" 
+
 
 	timer = timer + dtime
 
@@ -111,8 +111,15 @@ minetest.register_globalstep(function(dtime)
           minetest.chat_send_all("!!!Warning,Server will be restarted after 1 mins!!!")
           minetest.chat_send_all("!!!Server will be online again in a minute !!!")
 					recovery_md0_privs() --recovery supervisors's privs
-					os.execute("rm -rf "..check_file_worldedit)
-					os.execute("touch  "..check_file)
+					
+					if io.open(check_file_reset, "r") ~= nil then 
+						local check_file = minetest.get_worldpath() .. "/check_file" --means reset privs for supervisors
+						local check_file_worldedit = minetest.get_worldpath() .. "/check_file_worldedit" --means add worldedit privs for supervisors
+
+						os.execute("rm -rf "..check_file_worldedit)
+						os.execute("touch  "..check_file)
+					end
+					
     end
 		timer = 0
 	end
